Move spawn loop and related logic to scheduler.rs
Refactor to make scheduler.rs a true scheduler that handles execution:
- Move is_ready_to_spawn, ActorFn, ActorRegistry to scheduler.rs
- Move spawn_actor_task and spawn loop logic to run_spawn_loop()
- run_spawn_loop takes &RunHandle to access dag, suspension, actor_registry
- Environment.rs now focuses on build-phase configuration
- RunHandle::run() delegates to scheduler's run_spawn_loop()
