+++ ailets-rs/ailetos/src/scheduler.rs
@@ -1,7 +1,58 @@
-use std::collections::HashSet;
+//! Scheduler - execution scheduling for the actor system
+//!
+//! This module handles:
+//! - Topological ordering of DAG nodes
+//! - Readiness checking for node spawning
+//! - The spawn loop that runs actors
 
-use crate::dag::{Dag, NodeKind};
+use std::collections::{HashMap, HashSet};
+use std::sync::Arc;
+
+use tokio::sync::mpsc;
+use tracing::{debug, warn};
+
+use crate::dag::{Dag, NodeKind, NodeState};
+use crate::environment::RunHandle;
 use crate::idgen::Handle;
+use crate::pipe::PipePool;
+use crate::suspension::SuspensionState;
+use crate::system_runtime::IoRequest;
+use crate::{BlockingActorRuntime, KVBuffers, ShutdownHandle};
+
+/// Type for actor functions
+pub type ActorFn = fn(&dyn actor_runtime::ActorRuntime) -> Result<(), String>;
+
+/// Registry mapping actor names to their implementation functions
+#[derive(Clone)]
+pub struct ActorRegistry {
+    actors: HashMap<String, ActorFn>,
+}
+
+impl ActorRegistry {
+    #[must_use]
+    pub fn new() -> Self {
+        Self {
+            actors: HashMap::new(),
+        }
+    }
+
+    /// Register an actor function
+    pub fn register(&mut self, name: impl Into<String>, actor_fn: ActorFn) {
+        self.actors.insert(name.into(), actor_fn);
+    }
+
+    /// Get an actor function by name
+    #[must_use]
+    pub fn get(&self, name: &str) -> Option<ActorFn> {
+        self.actors.get(name).copied()
+    }
+}
+
+impl Default for ActorRegistry {
+    fn default() -> Self {
+        Self::new()
+    }
+}
 
 /// Conditions for stopping DAG iteration
 #[derive(Debug, Clone, Default)]
@@ -14,6 +65,168 @@ pub struct StopConditions {
     pub stop_after: Option<Handle>,
 }
 
+/// Decide whether a node is ready to be spawned.
+///
+/// Decision table (iterate concrete deps, first decisive result wins):
+///
+/// | Dep state              | Has output (pipe realized) | Decision       |
+/// |------------------------|----------------------------|----------------|
+/// | `NotStarted`           | —                          | don't start    |
+/// | Suspended              | —                          | don't start    |
+/// | Running / Terminating  | yes                        | start          |
+/// | Running / Terminating  | no                         | don't start    |
+/// | Terminated             | yes                        | start          |
+/// | Terminated             | no                         | skip (neutral) |
+/// | (all deps exhausted)   | —                          | start          |
+///
+/// "Has output" is checked optimistically: a dep has output if its stdout
+/// pipe is realized (writer exists in the pool), regardless of byte count.
+pub fn is_ready_to_spawn<K: KVBuffers>(
+    node_handle: Handle,
+    dag: &Dag,
+    pipe_pool: &PipePool<K>,
+    suspension: &SuspensionState,
+) -> bool {
+    use actor_runtime::StdHandle;
+
+    for dep in dag.resolve_dependencies(node_handle) {
+        let Some(dep_node) = dag.get_node(dep) else {
+            continue;
+        };
+
+        if suspension.is_suspended(dep) {
+            return false;
+        }
+
+        let has_output = pipe_pool
+            .get_already_realized_writer((dep, StdHandle::Stdout))
+            .is_some();
+
+        match dep_node.state {
+            NodeState::NotStarted => return false,
+            NodeState::Running | NodeState::Terminating => {
+                return has_output;
+            }
+            NodeState::Terminated => {
+                if has_output {
+                    return true;
+                }
+                // neutral: no output from this dep, continue to next
+            }
+        }
+    }
+
+    true
+}
+
+/// Spawn a task for an actor node
+fn spawn_actor_task(
+    node_handle: Handle,
+    idname: String,
+    actor_fn: ActorFn,
+    actor_runtime: BlockingActorRuntime,
+    shutdown: ShutdownHandle,
+) -> tokio::task::JoinHandle<()> {
+    tokio::task::spawn_blocking(move || {
+        debug!(node = ?node_handle, name = %idname, "task starting");
+
+        actor_runtime.register_std_fds();
+
+        let result = actor_fn(&actor_runtime);
+
+        match result {
+            Ok(()) => {
+                debug!(node = ?node_handle, name = %idname, "task completed");
+            }
+            Err(e) => {
+                warn!(node = ?node_handle, name = %idname, error = %e, "task error");
+            }
+        }
+
+        debug!(node = ?node_handle, name = %idname, "task done, shutdown via Drop");
+        drop(shutdown);
+    })
+}
+
+/// Run the spawn loop: spawn actors as their dependencies become ready
+///
+/// Returns the list of spawned actor tasks for the caller to await.
+pub async fn run_spawn_loop<K: KVBuffers>(
+    run_handle: &RunHandle<K>,
+    target: Handle,
+    stop_conditions: StopConditions,
+    pipe_pool: Arc<PipePool<K>>,
+    system_tx: mpsc::UnboundedSender<IoRequest>,
+    spawn_notify: Arc<tokio::sync::Notify>,
+) -> Vec<tokio::task::JoinHandle<()>> {
+    // Build initial pending list: NotStarted nodes in topological order
+    let mut pending: Vec<Handle> = {
+        let dag_guard = run_handle.dag.read();
+        let scheduler = Scheduler::with_stop_conditions(&dag_guard, target, stop_conditions);
+        scheduler
+            .iter()
+            .filter(|&n| {
+                dag_guard
+                    .get_node(n)
+                    .is_some_and(|node| node.state == NodeState::NotStarted)
+            })
+            .collect()
+    };
+
+    let mut actor_tasks = Vec::new();
+
+    // On-demand spawn loop: runs concurrently with actors
+    loop {
+        // Find all pending nodes whose inputs are ready
+        let to_spawn: Vec<(Handle, String)> = {
+            let dag_guard = run_handle.dag.read();
+            pending
+                .iter()
+                .filter(|&&n| is_ready_to_spawn(n, &dag_guard, &pipe_pool, &run_handle.suspension))
+                .filter_map(|&n| dag_guard.get_node(n).map(|node| (n, node.idname.clone())))
+                .collect()
+        };
+
+        // Spawn each ready node
+        for (node_handle, idname) in &to_spawn {
+            pending.retain(|&h| h != *node_handle);
+
+            debug!(node = ?node_handle, name = %idname, "spawning actor task");
+            run_handle.dag.write().set_state(*node_handle, NodeState::Running);
+
+            let (actor_runtime, shutdown) = BlockingActorRuntime::new(
+                *node_handle,
+                system_tx.clone(),
+                Arc::clone(&run_handle.suspension),
+            );
+
+            if let Some(actor_fn) = run_handle.actor_registry.get(idname) {
+                let task = spawn_actor_task(
+                    *node_handle,
+                    idname.clone(),
+                    actor_fn,
+                    actor_runtime,
+                    shutdown,
+                );
+                actor_tasks.push(task);
+            } else {
+                warn!(node = ?node_handle, name = %idname, "actor not registered, skipping");
+            }
+        }
+
+        // All nodes spawned — exit loop
+        if pending.is_empty() {
+            break;
+        }
+
+        // Wait for a pipe-realized or actor-terminated signal
+        spawn_notify.notified().await;
+    }
+
+    actor_tasks
+}
+
+/// Topological order iterator for DAG nodes
 pub struct Scheduler<'a> {
     dag: &'a Dag,
     target: Handle,
